added modules as docs
