--- docs/.gitignore
@@ -1,4 +1,0 @@
-build
-source/_autodoc
-docs.iml
-
