--- docs/requirements.txt
@@ -1,9 +1,0 @@
-myst-parser
-sphinx-rtd-theme
-autodoc
-sphinx>=4.4.0
-docutils>=0.17.1
-sphinx-tabs
-nbsphinx
-ipykernel
-raphtory
