+++ docs/source/_templates/pandas_footer.html
