+++ docs/source/api/index.rst
@@ -1,25 +1,16 @@
+{{ header }}
 
-Python
-=======
+.. _api:
 
-.. toctree::
-   :maxdepth: 3
-
-   python
-
-Rust
-=======
-
-*  `Rust Docs`_
-
-
-.. _Rust Docs: ../_static/raphtory/index.html
-
-
-Updating
-========
+=============
+API
+=============
 
 .. toctree::
-   :maxdepth: 3
+   :maxdepth: 2
 
-   update
+
+   raphtory
+   rust
+
+
