--- docs/source/api/python.rst
@@ -1,12 +1,0 @@
-API (Python)
-===
-
-.. autosummary::
-   :toctree: _autosummary
-
-   raphtory
-   raphtory.algorithms
-   raphtory.nullmodels
-   raphtory.vis
-   raphtory.graph_loader
-   raphtory.graph_gen
