+++ docs/source/api/raphtory.rst
@@ -1,0 +1,54 @@
+﻿.. _overview:
+
+{{ header }}
+
+****************
+Raphtory
+****************
+
+.. automodule:: raphtory
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :private-members:
+
+raphtory.vis
+-------------------
+
+.. automodule:: raphtory.vis
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :private-members:
+
+
+raphtory.nullmodels
+--------------------------
+
+.. automodule:: raphtory.nullmodels
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :private-members:
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
