+++ docs/source/api/rust.rst
@@ -1,0 +1,4 @@
+Rust
+================
+
+Our rust module is hosted on docs.rs, and can be found `here <https://docs.rs/raphtory/>`__.
