--- docs/source/install/rust/install.md
@@ -1,21 +1,0 @@
-# Install
-
-raphtory requires rust, please see the rust guide to [install](https://doc.rust-lang.org/stable/book/ch01-01-installation.html). 
-
-### 1. Set up raphtory 
-
-Clone the raphtory Repository
-
-```sh
-git clone git@github.com:Raphtory/raphtory.git
-```
-
-### 2. Build raphtory
-
-`cd`into the `raphtory` folder and run 
-
-```sh
-cargo build
-```
-
-Thats it! raphtory is now ready to use!
