--- docs/source/install/rust/test.md
@@ -1,6 +1,0 @@
-# Test
-
-- To run `raphtory` rust tests
-    ```bash
-    $ cargo test
-    ```
